feat(prettier): Forward all text to prettier
If we forward all text with it's filepath if there is one, prettier will figure out which parser it should use. This will make this package support all future parsers prettier till add in the future by default.

We only have to check which files to forward to eslint which is something that changes less frequently. For now we forward files with extension defined in eslint config or default to .js/.jsx/.ts/.tsx files.